refactoring2
