+++ modules/formatter/prompt_builder.py
@@ -1,5 +1,5 @@
-from .formatter import Formatter
-from .parser import parse_raw, check_special_tokens, parse_graphllm
+from .prompt_formatter import Formatter
+from .prompt_parser import parse_raw, check_special_tokens, parse_graphllm
 import re
 
 ## questa è la parte che non ha bisongno di conoscere il modello
